+++ AICoreD/Move.h
@@ -3,7 +3,7 @@
 struct Move
 {
 	int value;
-	short unsigned int row;
-	short unsigned int col;
-	short unsigned int target;
+	unsigned int row;
+	unsigned int col;
+	unsigned int target;
 };
